feat: Add working hours field to updateDriverByEmail function
The commit adds a new field for working hours in the `updateDriverByEmail` function of the `driver.action.ts` file. This field allows for updating the working hours of a driver, including the day, start time, and end time. This functionality will be used to manage and track the working hours of drivers in the application.